+++ src/mesido/workflows/grow_workflow.py
@@ -3,7 +3,6 @@ import logging
 import os
 import sys
 import time
-from typing import Dict
 
 from mesido.esdl.esdl_additional_vars_mixin import ESDLAdditionalVarsMixin
 from mesido.esdl.esdl_mixin import ESDLMixin
@@ -14,6 +13,7 @@ from mesido.workflows.io.write_output import ScenarioOutput
 from mesido.workflows.utils.adapt_profiles import (
     adapt_hourly_year_profile_to_day_averaged_with_hourly_peak_day,
 )
+from mesido.workflows.utils.error_types import HEAT_NETWORK_ERRORS, potential_error_to_error
 from mesido.workflows.utils.helpers import main_decorator, run_optimization_problem_solver
 
 import numpy as np
@@ -181,8 +181,6 @@ class EndScenarioSizing(
 
         self._save_json = False
 
-        self._asset_potential_errors = Dict[str, Dict]
-
     def parameters(self, ensemble_member):
         parameters = super().parameters(ensemble_member)
         parameters["peak_day_index"] = self.__indx_max_peak
@@ -202,31 +200,7 @@ class EndScenarioSizing(
         """
         super().read()
 
-        # Error checking:
-        # - installed capacity/power of a heating/cooling demand is sufficient for the specified
-        #   demand profile
-        is_error = False
-        for error_type, errors in self._asset_potential_errors.items():
-            if error_type in ["heat_demand.power", "cold_demand.power"]:
-                if len(errors) > 0:
-                    for asset_name in errors:
-                        logger.error(self._asset_potential_errors[error_type][asset_name])
-                    logger.error(
-                        "Asset insufficient installed capacity: please increase the"
-                        " installed power or reduce the demand profile peak value of the demand(s)"
-                        " listed."
-                    )
-                    is_error = True
-            elif error_type in ["heat_demand.type"]:
-                if len(errors) > 0:
-                    for asset_name in errors:
-                        logger.error(self._asset_potential_errors[error_type][asset_name])
-                    logger.error("Incorrect asset type: please update.")
-                    is_error = True
-
-        if is_error:
-            exit(1)
-        # end error checking
+        potential_error_to_error(HEAT_NETWORK_ERRORS)
 
         (
             self.__indx_max_peak,